requirements.txt
requests
argparse
pyyaml
colorama
rich
